Changes of InsertShipment Pull Request added: test will be added once 'weight missing' fix found
