+++ src/dto/general/Packages.php
@@ -4,31 +4,14 @@ namespace DeliveryMatchApiLibrary\dto\general;
 
 class Packages
 {
-    private ?string $warehouse;
-    private string $description;
-    private ?string $type;
-    private float $weight;
-    private float $length;
-    private float $width;
-    private float $height;
+    /** @var Package[] */
+    private array $packages;
 
     /**
-     * @param string|null $warehouse
-     * @param string $description
-     * @param string|null $type
-     * @param float $weight
-     * @param float $length
-     * @param float $width
-     * @param float $height
+     * @param Package[] $packages
      */
-    public function __construct(?string $warehouse, string $description, ?string $type, float $weight, float $length, float $width, float $height)
+    public function __construct(array $packages)
     {
-        $this->warehouse = $warehouse;
-        $this->description = $description;
-        $this->type = $type;
-        $this->weight = $weight;
-        $this->length = $length;
-        $this->width = $width;
-        $this->height = $height;
+        $this->packages = $packages;
     }
 }
