+++ src/dto/requests/InsertShipmentRequest.php
@@ -4,6 +4,7 @@ namespace DeliveryMatchApiLibrary\dto\requests;
 
 use DeliveryMatchApiLibrary\dto\general\Client;
 use DeliveryMatchApiLibrary\dto\general\Customer;
+use DeliveryMatchApiLibrary\dto\general\Package;
 use DeliveryMatchApiLibrary\dto\general\Packages;
 use DeliveryMatchApiLibrary\dto\general\Quote;
 use DeliveryMatchApiLibrary\dto\general\Sender;
@@ -15,7 +16,7 @@ class InsertShipmentRequest
     private Shipment $shipment;
     private ?Sender $sender;
     private Customer $customer;
-    private ?Packages $packages;
+    private ?Package $packages;
     private Quote $quote;
     private ?bool $fragileGoods;
     private ?bool $dangerousGoods;
@@ -28,7 +29,7 @@ class InsertShipmentRequest
      * @param Shipment $shipment
      * @param Sender|null $sender
      * @param Customer $customer
-     * @param Packages|null $packages
+     * @param Package|null $packages
      * @param Quote $quote
      * @param bool|null $fragileGoods
      * @param bool|null $dangerousGoods
@@ -36,7 +37,7 @@ class InsertShipmentRequest
      * @param float|null $priceExcl
      * @param float $weight
      */
-    public function __construct(Client $client, Shipment $shipment, ?Sender $sender, Customer $customer, ?Packages $packages, Quote $quote, ?bool $fragileGoods, ?bool $dangerousGoods, float $priceIncl, ?float $priceExcl, float $weight)
+    public function __construct(Client $client, Shipment $shipment, ?Sender $sender, Customer $customer, ?Package $packages, Quote $quote, ?bool $fragileGoods, ?bool $dangerousGoods, float $priceIncl, ?float $priceExcl, float $weight)
     {
         $this->client = $client;
         $this->shipment = $shipment;
